pyproject.toml
[tool.poetry]
name = "baldr"
version = "0.1.0"
description = ""
authors = ["Arnthor Jonsson <arnthorj2@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Pillow = "^8.4.0"
tqdm = "^4.62.3"

[tool.poetry.dev-dependencies]
black = "21.7b0"
isort = "^5.9.3"
mypy = "^0.931"
pre-commit = "^2.9.2"
pytest = "^5.2"
poetry2setup = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.mypy]
python_version = 3.7
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

